[ci skip] fix some DeprecatedWarning that will be FutureErrors

+++ python/pyhrf/vbjde/vem_tools.py
@@ -135,7 +135,7 @@ def compute_mat_X_2(nbscans, tr, lhrf, dt, onsets, durations=None):
 
     x = np.zeros((nbscans, lhrf), dtype=float)
     tmax = nbscans * tr  # total session duration
-    lgt = (nbscans + 2) * osf  # nb of scans if tr=dt
+    lgt = int((nbscans + 2) * osf)  # nb of scans if tr=dt
     paradigm_bins = restarize_events(onsets, np.zeros_like(onsets), dt, tmax)
     firstcol = np.concatenate(
         (paradigm_bins, np.zeros(lgt - len(paradigm_bins))))
